+++ async-magic.js
@@ -7,5 +7,6 @@ _api.promisifyAll = require('./lib/promisify-all');
 _api.wait = require('./lib/wait');
 _api.PromiseResolver = require('./lib/PromiseResolver');
 _api.parallel = require('./lib/parallel');
+_api.series = require('./lib/series');
 
 module.exports = _api;