+++ CHANGES.md
@@ -2,6 +2,7 @@
 * Added: `.wait` function as alias of [Promise.all](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Promise/all)
 * Added: `.PromiseResolver` to queue a promised function with given arguments for execution
 * Added: `.parallel` function to execute multiple `PromiseResolver` in parallel with given task limit
+* Added: `.series` function to execute multiple `PromiseResolver` in series
 
 ### 1.1.0 ###
 * Added: Mocha based Testcases